ActionService working

+++ src/Services/Commands/AddShaderCommand.hpp
@@ -1,18 +1,28 @@
 #include "Command.hpp"
-#include "../ShaderChainerService.hpp"
+#include "ShaderChainerService.hpp"
+#include "ShaderType.hpp"
 
 class AddShaderCommand : public Command {
 public:
-  AddShaderCommand(std::shared_ptr<Shader> shader) : shader(shader) {}
+  AddShaderCommand(ShaderType shaderType) : shaderType(shaderType) {}
 
   void execute() override {
+    if (!shader) {
+      shader = ShaderChainerService::getService()->makeShader(shaderType);
+    }
     ShaderChainerService::getService()->addShader(shader);
   }
 
   void undo() override {
-    ShaderChainerService::getService()->removeShader(shader);
+    if (shader) {
+      ShaderChainerService::getService()->removeShader(shader);
+    }
   }
 
-private:
+  std::shared_ptr<Shader> getShader() const {
+    return shader;
+  }
+
+  ShaderType shaderType;
   std::shared_ptr<Shader> shader;
 };
